Changes <osg/Notify> support to be more Pythonic
Exposing osg.NotifyHandler as a Python object isn't compatible
with OSG's "singleton" implementation; at least, not with a bunch
of ugly support code. It makes much more sense to just accept a
generic "callable" as the notification handler, and we'll likely
develop a more generic wrapper for handling singletons in the
future (that work not only for notification, but wherever else we
encounter singletons with annoying static data).