Introduces HUGE restructuring changes
This patch changes source/header file names to be more consistent
with their Python-centric behavior (and less ambiguous when used
with other OSG projects). Furthmore, a common static library
containing ALL the wrappers/bindings is now always compiled
(libOpenSceneGraph_python.a on Linux), and then linked against
either the typical CPython module OR an embedded target. Moving
forward, special attention will need to be paid to when/where the
PYOSG_INTERNAL macro is necessary (and how this will be addressed
when MSVC support is added).