Adds a VERY THIN ArgumentParser wrapper
There's NO REASON for people to ACTUALLY NEED to use the argument
parsing object from OSG when the entire process is so much easier
in Python itself. This patch adds a thin construction-only wrapper
for `osg::ArgumentParser`, only intended to be used for those APIs
that expect a properly-created instance of one (which is rare).